Several changes
Added lua/emit.lua which is a class for building emit tables. Therefore,
the Vis table is low-level, the VisUtil is mid-level, and the Emit is
high-level for emitting particles.

Added a SHIFT parameter to the Lua keydown and keyup callbacks.

Made various modifications to VisUtil

Added Vis.emitnow(Vis.script, ...)

Added test/9_mouse.lua to demonstrate interactivity more