Significant changes to the engine and utilities
Removed the need for flipping the images when rendering to files; that's now done automatically inside the drawer. Removed all scripting related to that.

Added a frameskip argument to vis, -s. It doesn't change the flist; rather, the drawer just doesn't save those first few frames to files, as if they never happened. They're still executed though; the frameskip is just for the drawer saving files to pngs.

Changed Vis.audio(filename) to Vis.audio(Vis.flist, when, filename) to fix a problem where the audio would start while the script is still running. The longer the script took to run, the worse the desync. With this in place, the desync went from 10 frames (3rd of a second) to just four frames (2/15ths of a second). The work-around is in the Makefile.

Almost rewrote the entire encoding logic for the scripts/process.py script. Removed all references to ImageMagick and allowed paths to have options that modify them, like ["-i", "media/Bowser.wav::-s 00:00:00.30"] to skip the first 300ms of Bowser.wav. It works for input, output, and -i values.

Added a PLAY ftype to implement the Vis.audio change.