Added -L <script> argument: run script verbatim
Example: ./vis -L "print(VisUtil.strobject(_G))"

Added a bit more debugging and cleaned up the code significantly.

Added support for a semi-transparent "pixel" image (pixel_3x3.png) which makes the emit look significantly nicer. There are some problems with it still but those can be worked out. It's not possible to disable this (short of removing the file) but that'll be an option eventually.

Added a third argument to Vis.onmousedown and Vis.onmouseup: the mouse button pressed.
    1 = left mouse button
    2 = middle mouse button
    3 = right mouse button
    4 = extra button 1 (only present on certain mice)
    5 = extra button 2 (only present on certain mice)

And improved VisUtil.strobject to be more informative when seeing duplicate tables.