Scrapped the idea for executing >1 script
You can pass the option -l as many times as you want (as it always has been the case), but only the last one will be used (again, it was always the case).

I tried to get every script passed via -l to be executed, but this tured out to be far too much work than I anticipated.

The klist has no use anymore. It's there for convenience should I need it later, but if that never comes I'll probably remove it.