Added a way to pass values from the command line to Lua: either by using the -A command line option or by specifying -- on the command line, which fowards all subsequent values to Lua. The values are stored in a global array named 'Arguments'. The -A argument can be passed multiple times.
Better handling of unrecognized arguments: they're printed after parsing.

Added VisUtil.Debug, which is meant to act like printf:
    VisUtil.Debug("There was an error: %d: %s", the_error, the_error_str)
or as a way to inspect objects:
    VisUtil.Debug(_G) -- causes the table _G to be formatted and printed
or a number of other ways.

Significant code cleanup and code improvement in script.c. The Arguments global is built every time the Lua runtime is executed, which is probably not the best of ideas.

Fixed a nasty bug in test/test_5_emitcls.lua caused by me defining a function called "debug" in the same scope as the debug module. Whoops.